(auth_stat) split into its own file
Matching clnt_stat
